requirements.txt
kafka-python
requests
influxdb
